pyproject.toml
[project]
name = "distance-preservation-encoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.0.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.0",
    "torch>=2.7.1",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu126" },
]
torchvision = [
    { index = "pytorch-cu126" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
